Fine frequency correction is now exact
Fine frequency correction was rounded to 1 Hz before.
Update dabstar.pro

+++ src/base/ofdm/sample_reader.h
@@ -59,8 +59,8 @@ public:
   bool is_running() const { return running.load(); }
   void discard_samples(i32 iSampleCnt);
   void get_linear_peak_level_and_clear(f32 & oLevelPeak, f32 & oLevelMean);
-  cf32 get_sample(i32);
-  void get_samples(TArrayTn & oV, const i32 iStartIdx, i32 iNoSamples, const i32 iFreqOffsetBBHz, bool iShowSpec);
+  cf32 get_sample(f32);
+  void get_samples(TArrayTn & oV, const i32 iStartIdx, i32 iNoSamples, const f32 iFreqOffsetBBHz, bool iShowSpec);
   void start_dumping(SNDFILE *);
   void stop_dumping();
   void set_dc_and_iq_correction(bool iDoDcCorr, bool iDoIqCorr);
@@ -116,4 +116,3 @@ signals:
   void signal_show_spectrum(i32);
   void signal_show_cir(i32);
 };
-
